+++ testData/org/elixir_lang/heex/reference/component_find_usages/default_arity/page_live.ex
@@ -1,0 +1,5 @@
+defmodule MyAppWeb.PageLive do
+  def but<caret>ton(assigns \\ %{}) do
+    ~H""
+  end
+end